% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cexportuser-package.R
\docType{package}
\name{cexportuser-package}
\alias{cexportuser}
\alias{cexportuser-package}
\title{cexportuser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/cexportuser}
  \item Report bugs at \url{https://github.com/DavisVaughan/cexportuser/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
